# flutter_application_1/1XL-Infra-Office-Asset-Tracking-Maintenance-System-main (1)/1XL-Infra-Office-Asset-Tracking-Maintenance-System-main/server/openapi.yaml
openapi: 3.1.0
info:
  title: 1XL Asset Tracker — Mobile API
  version: 1.0.0
  description: |
    REST API consumed by the Flutter companion app for the 1XL Asset Tracker
    platform. Implements every P0 endpoint from `questionnaire.md` Section 5,
    plus photo upload, recovery, audit verification, and reference lookups.

    ### Auth model
    All requests use the standard `Authorization: Bearer <token>` scheme.
    The bearer token is one of:
    - **API key** — `<orgslug>_<live|test>_<24chars>`, e.g. `1xl_live_K8j3pQrX9bR2sN7wHzVc1qLm`.
      Issued by an org admin in the web UI.
    - **Access JWT** — short-lived per-user token from `POST /auth/login`.
    The server distinguishes them by shape. `/auth/login`, `/auth/refresh`,
    `/auth/logout` only accept an API key. `/auth/me` only accepts a user JWT.
    Everything else accepts either.

    ### Versioning
    Additive changes within `/api/v1/*`. Breaking changes get `/api/v2/*`.

    ### Idempotency
    All `POST` endpoints accept `Idempotency-Key`. Same key + same body = cached
    response for 24h. Same key + different body = `409 CONFLICT`.
servers:
  - url: https://asset.thebizzfly.com
    description: Production
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>` where `<token>` is either:
          - an API key (`<orgslug>_<live|test>_<24chars>`) — for /auth/login/refresh/logout and server-to-server calls, or
          - a user JWT (HS256) — for user-scoped calls (e.g. /auth/me).
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, minLength: 16, maxLength: 128 }
      description: Required for safe POST retries. 24h TTL.
    AssetId:
      name: assetId
      in: path
      required: true
      schema: { type: string, format: uuid }
    RepairId:
      name: repairId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - UNAUTHENTICATED
                - TOKEN_EXPIRED
                - INVALID_API_KEY
                - FORBIDDEN
                - INSUFFICIENT_SCOPE
                - NOT_FOUND
                - ASSET_NOT_FOUND
                - REPAIR_NOT_FOUND
                - PHOTO_NOT_FOUND
                - CYCLE_NOT_FOUND
                - INVALID_QR_PAYLOAD
                - VALIDATION_FAILED
                - CONFLICT
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:    { type: string }
            requestId:  { type: string }
            fieldErrors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:   { type: string }
                  message: { type: string }
            details: { }
    User:
      type: object
      properties:
        id:             { type: string, format: uuid }
        name:           { type: string }
        email:          { type: string, format: email }
        role:           { type: string, enum: [admin, manager, employee, technician, vendor, auditor, staff] }
        departmentId:   { type: string, nullable: true }
        phone:          { type: string }
        avatar:         { type: string, nullable: true }
        organizationId: { type: string, format: uuid, nullable: true }
        isActive:       { type: boolean }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Conflict:
      description: Idempotency-key reuse with different body, or business rule conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
paths:
...
  /api/v1/assets/{assetId}/repairs:
    get:
      summary: List repair tickets for an asset
      parameters: [{ $ref: '#/components/parameters/AssetId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { type: object } }
                  total: { type: integer }
    post:
      summary: Raise a repair ticket
      parameters:
        - { $ref: '#/components/parameters/AssetId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issue]
              properties:
                issue:        { type: string, minLength: 3 }
                priority:     { type: string, enum: [low, medium, high, critical] }
                vendorId:     { type: string, format: uuid, nullable: true }
                technicianId: { type: string, format: uuid, nullable: true }
                partsUsed:    { type: string }
                notes:        { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  repair: { type: object }
  /api/v1/repairs/{repairId}:
    get:
      summary: Full repair ticket with vendor + update history
      parameters: [{ $ref: '#/components/parameters/RepairId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  repair:  { type: object }
                  vendor:  { type: object, nullable: true }
                  updates: { type: array, items: { type: object } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/repairs/{repairId}/updates:
    post:
      summary: Append an update to a repair ticket (status transition + note)
      parameters:
        - { $ref: '#/components/parameters/RepairId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [note]
              properties:
                statusTo:   { type: string, enum: [pending, assigned, in_progress, completed, cancelled] }
                note:       { type: string, minLength: 1 }
                partsUsed:  { type: string }
                laborHours: { type: number, minimum: 0 }
                cost:       { type: number, minimum: 0 }
      responses:
        '201':
          description: Update appended
          content:
            application/json:
              schema:
                type: object
                properties:
                  update: { type: object }

  # ─── Recovery (atomic per TRD 5.16) ───────────────────────────────────────
  /api/v1/assets/{assetId}/recovery:
    post:
      summary: File a loss/damage/theft incident, optionally marking the asset dead atomically
      parameters:
        - { $ref: '#/components/parameters/AssetId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [incidentType, description]
              properties:
                incidentType:  { type: string, enum: [lost, damaged, stolen, insurance_claim, write_off] }
                severity:      { type: string, enum: [low, medium, high, critical] }
                description:   { type: string, minLength: 3 }
                estimatedLoss: { type: number, minimum: 0 }
                incidentDate:  { type: string, format: date-time }
                markAssetDead: { type: boolean, description: 'When true, also flips the asset to status=dead. Atomic — recovery is rolled back on failure.' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  recovery: { type: object }
                  asset:    { type: object, nullable: true, description: 'Updated asset row when markAssetDead was true.' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ─── Audit ────────────────────────────────────────────────────────────────
...
  /api/v1/reference/users:
    get:
      summary: Picker list of users in the caller's org
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
                  total: { type: integer }
  /api/v1/reference/locations:
    get:
      summary: Picker list of locations in the caller's org
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { type: object }
                  total: { type: integer }